day 5 tdd style: finished
So I thought I'd try to do the day 5 puzzle in a very strict
tdd way and see how long it takes. Along the way, I made a commit
every time I had a new small step working, so I think that series of
commits is the most interesting thing here.

Overall, it took me about 90 minutes as compared to just 40 minutes
when I did the actual writing of the my original solution that I used
to finish this puzzle.

But doesn't that mean TDD is terrible? I don't think so. Why?

- During the tdd path I had *far* fewer broad confusing moments
  and I always had a good picture of what was working so far.

- Little mistakes that would have bitten me later were caught
  right away.

- I probably wrote more tests than I might have otherwise because
  I was trying to be strict in my use of tdd.

- Like a lot of design techniques, they have overheads which are
  most obvious on very small problems like this one. In a more
  realistic size of program, I think the benefits of tdd start to
  show though *very* quickly, especially as you're discovering the
  design instead of knowing it ahead of time like I did when
  I started this second pass.